+++ jest.config.js
@@ -6,6 +6,10 @@
 module.exports = {
   // All test configuration should be picked up by individual jest.config.js.
   // Jest has some issue when running in multiple projects with different coverage exclusion.
-  projects: ['<rootDir>/generator', '<rootDir>/packages/fhir-core'],
+  projects: [
+    '<rootDir>/app/generator/jest.config.js',
+    '<rootDir>/app/generator-datamodel-ftest/jest.config.js',
+    '<rootDir>/packages/fhir-core/jest.config.js',
+  ],
   coverageDirectory: '<rootDir>/coverage',
 };